requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-filter==2.1.0
django-tenant-schemas==1.9.0
djangorestframework==3.9.2
idna==2.8
Markdown==3.1
psycopg2==2.7.3.2
pytz==2019.1
requests==2.21.0
soupsieve==1.9.1
sqlparse==0.3.0
urllib3==1.24.2
